server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pocketbase>=0.15.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "tqdm>=4.67.3",
]
